move comps inside src
