+++ apps/landing/src/app/components/Footer/index.tsx
